# webui/frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Strategy Research API",
    "description": "HTTP API for quantnodes strategy research framework",
    "version": "0.3.0"
  },
  "paths": {
...
    "/api/hypothesis/create": {
      "post": {
        "tags": [
          "hypothesis"
        ],
        "summary": "Hypothesis Create",
        "description": "创建新 hypothesis。",
        "operationId": "hypothesis_create_api_hypothesis_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HypothesisCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/hypothesis/list": {
      "get": {
        "tags": [
          "hypothesis"
        ],
        "summary": "Hypothesis List",
        "description": "列出 hypotheses。",
        "operationId": "hypothesis_list_api_hypothesis_list_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/hypothesis/search": {
      "get": {
        "tags": [
          "hypothesis"
        ],
        "summary": "Hypothesis Search",
        "description": "搜索 hypotheses。",
        "operationId": "hypothesis_search_api_hypothesis_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/hypothesis/{hypothesis_id}": {
      "get": {
        "tags": [
          "hypothesis"
        ],
        "summary": "Hypothesis Get",
        "description": "获取单个 hypothesis。",
        "operationId": "hypothesis_get_api_hypothesis__hypothesis_id__get",
        "parameters": [
          {
            "name": "hypothesis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Hypothesis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/hypothesis/update": {
      "put": {
        "tags": [
          "hypothesis"
        ],
        "summary": "Hypothesis Update",
        "description": "更新 hypothesis。",
        "operationId": "hypothesis_update_api_hypothesis_update_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HypothesisUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HypothesisCreateRequest": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "thesis": {
            "type": "string",
            "title": "Thesis",
            "default": ""
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "exploring"
          },
          "universe": {
            "type": "string",
            "title": "Universe",
            "default": ""
          },
          "signal_definition": {
            "type": "string",
            "title": "Signal Definition",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "HypothesisCreateRequest"
      },
      "HypothesisUpdateRequest": {
        "properties": {
          "hypothesis_id": {
            "type": "string",
            "title": "Hypothesis Id"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "thesis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thesis"
          },
          "conclusion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conclusion"
          }
        },
        "type": "object",
        "required": [
          "hypothesis_id"
        ],
        "title": "HypothesisUpdateRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}